Move activation utility code to its own file
