+++ src/omnicore/omnicore.cpp
@@ -6,6 +6,7 @@
 
 #include "omnicore/omnicore.h"
 
+#include "omnicore/activation.h"
 #include "omnicore/convert.h"
 #include "omnicore/dex.h"
 #include "omnicore/encoding.h"
