+++ src/omnicore/rpc.cpp
@@ -6,6 +6,7 @@
 
 #include "omnicore/rpc.h"
 
+#include "omnicore/activation.h"
 #include "omnicore/convert.h"
 #include "omnicore/dex.h"
 #include "omnicore/errors.h"
