+++ src/omnicore/rules.cpp
@@ -6,6 +6,7 @@
 
 #include "omnicore/rules.h"
 
+#include "omnicore/activation.h"
 #include "omnicore/log.h"
 #include "omnicore/omnicore.h"
 #include "omnicore/notifications.h"
@@ -28,16 +29,6 @@
 
 namespace mastercore
 {
-/**
- * Pending activations
- */
-std::vector<FeatureActivation> vecPendingActivations;
-
-/**
- * Completed activations
- */
-std::vector<FeatureActivation> vecCompletedActivations;
-
 /**
  * Returns a mapping of transaction types, and the blocks at which they are enabled.
  */
@@ -323,138 +314,6 @@ bool IsAllowedOutputType(int whichType, int nBlock)
     }
 
     return false;
-}
-
-/**
- * Adds a feature activation to the PendingActivations vector.
- *
- */
-void AddPendingActivation(uint16_t featureId, int activationBlock, uint32_t minClientVersion, const std::string& featureName)
-{
-    FeatureActivation featureActivation;
-
-    featureActivation.featureId = featureId;
-    featureActivation.featureName = featureName;
-    featureActivation.activationBlock = activationBlock;
-    featureActivation.minClientVersion = minClientVersion;
-
-    vecPendingActivations.push_back(featureActivation);
-}
-
-/**
- * Deletes a pending activation from the PendingActivations vector.  Deletion is not supported on the CompletedActivations vector.
- *
- */
-void DeletePendingActivation(uint16_t featureId)
-{
-   for (std::vector<FeatureActivation>::iterator it = vecPendingActivations.begin(); it != vecPendingActivations.end(); ) {
-       if ((*it).featureId == featureId) {
-           it = vecPendingActivations.erase(it);
-       } else {
-           it++;
-       }
-   }
-}
-
-/**
- * Checks if any activations went live in the block
- *
- */
-void CheckLiveActivations(int blockHeight)
-{
-    std::vector<FeatureActivation> vecPendingActivations = GetPendingActivations();
-    for (std::vector<FeatureActivation>::iterator it = vecPendingActivations.begin(); it != vecPendingActivations.end(); ++it) {
-       if ((*it).activationBlock > blockHeight) continue;
-       FeatureActivation liveActivation = *it;
-       if (OMNICORE_VERSION < liveActivation.minClientVersion) {
-           std::string msgText = strprintf("Shutting down due to unsupported feature activation (%d: %s)", liveActivation.featureId, liveActivation.featureName);
-           PrintToLog(msgText);
-           PrintToConsole(msgText);
-           if (!GetBoolArg("-overrideforcedshutdown", false)) {
-               AbortNode(msgText, msgText);
-           }
-       }
-       PendingActivationCompleted(liveActivation.featureId);
-       uiInterface.OmniStateChanged();
-    }
-}
-
-/**
- * Moves an activation from PendingActivations to CompletedActivations when it is activated
- *
- */
-void PendingActivationCompleted(uint16_t featureId)
-{
-   for (std::vector<FeatureActivation>::iterator it = vecPendingActivations.begin(); it != vecPendingActivations.end(); ) {
-       if ((*it).featureId == featureId) {
-           FeatureActivation completedActivation = *it;
-           vecCompletedActivations.push_back(completedActivation);
-           it = vecPendingActivations.erase(it);
-       } else {
-           it++;
-       }
-   }
-}
-
-/**
- * Returns the vector of pending activations.
- */
-std::vector<FeatureActivation> GetPendingActivations()
-{
-    return vecPendingActivations;
-}
-
-/**
- * Returns the vector of completed activations.
- */
-std::vector<FeatureActivation> GetCompletedActivations()
-{
-    return vecCompletedActivations;
-}
-
-/**
- * TODO: currently clone of CheckAlertAuthorization - worth merging the functions?
- *
- * Determines whether the sender is an authorized source for Omni Core feature activation.
- *
- * The option "-omniactivationallowsender=source" can be used to whitelist additional sources,
- * and the option "-omniactivationignoresender=source" can be used to ignore a source.
- *
- * To consider any activation as authorized, "-omniactivationallowsender=any" can be used. This
- * should only be done for testing purposes!
- */
-bool CheckActivationAuthorization(const std::string& sender)
-{
-    std::set<std::string> whitelisted;
-
-    // Mainnet
-    whitelisted.insert("1OmniFoundation1111111111111111111"); // Omni Foundation <????@omni.foundation>
-
-    // Testnet / Regtest
-    // use -omniactivationallowsender for testing
-
-    // Add manually whitelisted sources
-    if (mapArgs.count("-omniactivationallowsender")) {
-        const std::vector<std::string>& sources = mapMultiArgs["-omniactivationallowsender"];
-
-        for (std::vector<std::string>::const_iterator it = sources.begin(); it != sources.end(); ++it) {
-            whitelisted.insert(*it);
-        }
-    }
-
-    // Remove manually ignored sources
-    if (mapArgs.count("-omniactivationignoresender")) {
-        const std::vector<std::string>& sources = mapMultiArgs["-omniactivationignoresender"];
-
-        for (std::vector<std::string>::const_iterator it = sources.begin(); it != sources.end(); ++it) {
-            whitelisted.erase(*it);
-        }
-    }
-
-    bool fAuthorized = (whitelisted.count(sender) ||
-                        whitelisted.count("any"));
-
-    return fAuthorized;
 }
 
 /**
